files added, updated and removed
the unneeded docs were removed such as the project guidance readme
a few files were renamed, and a README file recreated, though still a work in progress.
an image has been added to the image folder in order to have a banner image on the readme.
accessibility testing document added to the asset folder.